refactor(engine): move tests into more specific test class
Created two more classes in order to properly test form events applier
for creation/deletion
