+++ engine/src/test/java/io/camunda/zeebe/engine/state/appliers/FormCreatedApplierTest.java
@@ -5,43 +5,85 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.deployment;
+package io.camunda.zeebe.engine.state.appliers;
 
+import static io.camunda.zeebe.util.buffer.BufferUtil.bufferAsArray;
 import static io.camunda.zeebe.util.buffer.BufferUtil.bufferAsString;
 import static io.camunda.zeebe.util.buffer.BufferUtil.wrapString;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.camunda.zeebe.engine.state.appliers.FormCreatedApplier;
-import io.camunda.zeebe.engine.state.appliers.FormDeletedApplier;
+import io.camunda.zeebe.engine.state.deployment.PersistedForm;
 import io.camunda.zeebe.engine.state.mutable.MutableFormState;
 import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
-import io.camunda.zeebe.engine.util.ProcessingStateRule;
+import io.camunda.zeebe.engine.util.ProcessingStateExtension;
 import io.camunda.zeebe.protocol.impl.record.value.deployment.FormRecord;
 import io.camunda.zeebe.stream.api.state.KeyGenerator;
 import io.camunda.zeebe.test.util.Strings;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-public class FormStateMultiTenantTest {
+@ExtendWith(ProcessingStateExtension.class)
+public class FormCreatedApplierTest {
 
   private static final String TENANT_1 = "tenant1";
   private static final String TENANT_2 = "tenant2";
-  @Rule public final ProcessingStateRule stateRule = new ProcessingStateRule();
-
-  private MutableFormState formState;
-  private MutableProcessingState processingState;
   private KeyGenerator keyGenerator;
+  private MutableProcessingState processingState;
+  private MutableFormState formState;
   private FormCreatedApplier formCreatedApplier;
-  private FormDeletedApplier formDeletedApplier;
 
-  @Before
-  public void setUp() {
-    processingState = stateRule.getProcessingState();
+  @BeforeEach
+  public void setup() {
     formState = processingState.getFormState();
-    keyGenerator = processingState.getKeyGenerator();
     formCreatedApplier = new FormCreatedApplier(formState);
-    formDeletedApplier = new FormDeletedApplier(formState);
+    keyGenerator = processingState.getKeyGenerator();
+  }
+
+  @Test
+  void shouldStoreForm() {
+    // given
+    final var formRecord = sampleFormRecord();
+    formCreatedApplier.applyState(formRecord.getFormKey(), formRecord);
+
+    // when
+    final var maybePersistedForm = formState.findFormByKey(formRecord.getFormKey(), TENANT_1);
+
+    // then
+    assertThat(maybePersistedForm).isNotEmpty();
+    final var persistedForm = maybePersistedForm.get();
+    assertThat(bufferAsString(persistedForm.getFormId())).isEqualTo(formRecord.getFormId());
+    assertThat(persistedForm.getVersion()).isEqualTo(formRecord.getVersion());
+    assertThat(persistedForm.getFormKey()).isEqualTo(formRecord.getFormKey());
+    assertThat(bufferAsString(persistedForm.getResourceName()))
+        .isEqualTo(formRecord.getResourceName());
+    assertThat(bufferAsArray(persistedForm.getChecksum())).isEqualTo(formRecord.getChecksum());
+    assertThat(bufferAsArray(persistedForm.getResource())).isEqualTo(formRecord.getResource());
+  }
+
+  @Test
+  void shouldFindLatestByFormId() {
+    // given
+    final var formRecord1 = sampleFormRecord();
+    formCreatedApplier.applyState(formRecord1.getFormKey(), formRecord1);
+
+    final var formRecord2 = sampleFormRecord(2L, "form-id", 2, TENANT_1);
+    formCreatedApplier.applyState(formRecord2.getFormKey(), formRecord2);
+
+    // when
+    final var maybePersistedForm =
+        formState.findLatestFormById(formRecord1.getFormIdBuffer(), TENANT_1);
+
+    // then
+    assertThat(maybePersistedForm).isNotEmpty();
+    final var persistedForm = maybePersistedForm.get();
+    assertThat(bufferAsString(persistedForm.getFormId())).isEqualTo(formRecord2.getFormId());
+    assertThat(persistedForm.getVersion()).isEqualTo(2);
+    assertThat(persistedForm.getFormKey()).isEqualTo(formRecord2.getFormKey());
+    assertThat(bufferAsString(persistedForm.getResourceName()))
+        .isEqualTo(formRecord2.getResourceName());
+    assertThat(bufferAsArray(persistedForm.getChecksum())).isEqualTo(formRecord2.getChecksum());
+    assertThat(bufferAsArray(persistedForm.getResource())).isEqualTo(formRecord2.getResource());
   }
 
   @Test
@@ -50,8 +92,8 @@ public class FormStateMultiTenantTest {
     final long formKey = keyGenerator.nextKey();
     final String formId = Strings.newRandomValidBpmnId();
     final int version = 1;
-    final var tenant1Form = createFormRecord(formKey, formId, version, TENANT_1);
-    final var tenant2Form = createFormRecord(formKey, formId, version, TENANT_2);
+    final var tenant1Form = sampleFormRecord(formKey, formId, version, TENANT_1);
+    final var tenant2Form = sampleFormRecord(formKey, formId, version, TENANT_2);
 
     // when
     formCreatedApplier.applyState(tenant1Form.getFormKey(), tenant1Form);
@@ -69,27 +111,11 @@ public class FormStateMultiTenantTest {
     assertPersistedForm(form2, formKey, formId, version, TENANT_2);
   }
 
-  @Test
-  public void shouldDeleteFormForSpecificTenant() {
-    // given
-    final long formKey = keyGenerator.nextKey();
-    final String formId = Strings.newRandomValidBpmnId();
-    final int version = 1;
-    final var tenant1Form = createFormRecord(formKey, formId, version, TENANT_1);
-    final var tenant2Form = createFormRecord(formKey, formId, version, TENANT_2);
-
-    formCreatedApplier.applyState(tenant1Form.getFormKey(), tenant1Form);
-    formCreatedApplier.applyState(tenant2Form.getFormKey(), tenant2Form);
-
-    // when
-    formDeletedApplier.applyState(tenant1Form.getFormKey(), tenant1Form);
-
-    // then
-    assertThat(formState.findLatestFormById(tenant1Form.getFormIdBuffer(), TENANT_1)).isEmpty();
-    assertThat(formState.findLatestFormById(tenant2Form.getFormIdBuffer(), TENANT_2)).isNotEmpty();
+  private FormRecord sampleFormRecord() {
+    return sampleFormRecord(1L, "form-id", 1, TENANT_1);
   }
 
-  private FormRecord createFormRecord(
+  private FormRecord sampleFormRecord(
       final long key, final String id, final int version, final String tenant) {
     return new FormRecord()
         .setFormKey(key)
