+++ engine/src/test/java/io/camunda/zeebe/engine/state/deployment/FormStateTest.java
@@ -28,62 +28,11 @@ public class FormStateTest {
   private final String tenantId = "<default>";
   private MutableProcessingState processingState;
   private MutableFormState formState;
-  private FormCreatedApplier formCreatedApplier;
-  private FormDeletedApplier formDeletedApplier;
 
   @BeforeEach
   public void setup() {
     formState = processingState.getFormState();
-    formCreatedApplier = new FormCreatedApplier(formState);
-    formDeletedApplier = new FormDeletedApplier(formState);
   }
-
-  @Test
-  void shouldStoreForm() {
-    // given
-    final var formRecord = sampleFormRecord();
-    formCreatedApplier.applyState(formRecord.getFormKey(), formRecord);
-
-    // when
-    final var maybePersistedForm = formState.findFormByKey(formRecord.getFormKey(), tenantId);
-
-    // then
-    assertThat(maybePersistedForm).isNotEmpty();
-    final var persistedForm = maybePersistedForm.get();
-    assertThat(bufferAsString(persistedForm.getFormId())).isEqualTo(formRecord.getFormId());
-    assertThat(persistedForm.getVersion()).isEqualTo(formRecord.getVersion());
-    assertThat(persistedForm.getFormKey()).isEqualTo(formRecord.getFormKey());
-    assertThat(bufferAsString(persistedForm.getResourceName()))
-        .isEqualTo(formRecord.getResourceName());
-    assertThat(bufferAsArray(persistedForm.getChecksum())).isEqualTo(formRecord.getChecksum());
-    assertThat(bufferAsArray(persistedForm.getResource())).isEqualTo(formRecord.getResource());
-  }
-
-  @Test
-  void shouldFindLatestByFormId() {
-    // given
-    final var formRecord1 = sampleFormRecord();
-    formCreatedApplier.applyState(formRecord1.getFormKey(), formRecord1);
-
-    final var formRecord2 = sampleFormRecord(2, 2L);
-    formCreatedApplier.applyState(formRecord2.getFormKey(), formRecord2);
-
-    // when
-    final var maybePersistedForm =
-        formState.findLatestFormById(formRecord1.getFormIdBuffer(), tenantId);
-
-    // then
-    assertThat(maybePersistedForm).isNotEmpty();
-    final var persistedForm = maybePersistedForm.get();
-    assertThat(bufferAsString(persistedForm.getFormId())).isEqualTo(formRecord2.getFormId());
-    assertThat(persistedForm.getVersion()).isEqualTo(2);
-    assertThat(persistedForm.getFormKey()).isEqualTo(formRecord2.getFormKey());
-    assertThat(bufferAsString(persistedForm.getResourceName()))
-        .isEqualTo(formRecord2.getResourceName());
-    assertThat(bufferAsArray(persistedForm.getChecksum())).isEqualTo(formRecord2.getChecksum());
-    assertThat(bufferAsArray(persistedForm.getResource())).isEqualTo(formRecord2.getResource());
-  }
-
   @Test
   void shouldReturnEmptyIfNoFormIsDeployedForFormId() {
     // when
@@ -100,100 +49,5 @@ public class FormStateTest {
 
     // then
     assertThat(persistedForm).isEmpty();
-  }
-
-  @Test
-  void shouldNotFindFormAfterDeletion() {
-    // given
-    final var form = sampleFormRecord();
-    formCreatedApplier.applyState(form.getFormKey(), form);
-
-    assertThat(formState.findLatestFormById(form.getFormIdBuffer(), form.getTenantId()))
-        .isNotEmpty();
-
-    // when
-    formDeletedApplier.applyState(form.getFormKey(), form);
-
-    // then
-    assertThat(formState.findLatestFormById(form.getFormIdBuffer(), form.getTenantId())).isEmpty();
-  }
-
-  @Test
-  void shouldNotFindVersion2AsLatestFormAfterDeletion() {
-    // given
-    final var formV1 = sampleFormRecord();
-    final var formV2 = sampleFormRecord(2, 2L);
-    formCreatedApplier.applyState(formV1.getFormKey(), formV1);
-    formCreatedApplier.applyState(formV2.getFormKey(), formV2);
-
-    final var latestFormOpt =
-        formState.findLatestFormById(formV2.getFormIdBuffer(), formV2.getTenantId());
-    assertThat(latestFormOpt).isNotEmpty();
-    assertThat(latestFormOpt.get().getVersion()).isEqualTo(2);
-
-    // when
-    formDeletedApplier.applyState(formV2.getFormKey(), formV2);
-
-    // then
-    final var latestFormV1Opt =
-        formState.findLatestFormById(formV2.getFormIdBuffer(), formV2.getTenantId());
-    assertThat(latestFormV1Opt).isNotEmpty();
-    assertThat(latestFormV1Opt.get().getVersion()).isEqualTo(1);
-  }
-
-  @Test
-  void shouldFindVersion2AsLatestFormAfterDeletion() {
-    // given
-    final var formV1 = sampleFormRecord();
-    final var formV2 = sampleFormRecord(2, 2L);
-    formCreatedApplier.applyState(formV1.getFormKey(), formV1);
-    formCreatedApplier.applyState(formV2.getFormKey(), formV2);
-
-    final var latestFormOpt =
-        formState.findLatestFormById(formV2.getFormIdBuffer(), formV2.getTenantId());
-    assertThat(latestFormOpt).isNotEmpty();
-    assertThat(latestFormOpt.get().getVersion()).isEqualTo(2);
-
-    // when
-    formDeletedApplier.applyState(formV1.getFormKey(), formV1);
-
-    // then
-    final var latestFormV2Opt =
-        formState.findLatestFormById(formV2.getFormIdBuffer(), formV2.getTenantId());
-    assertThat(latestFormV2Opt).isNotEmpty();
-    assertThat(latestFormV2Opt.get().getVersion()).isEqualTo(2);
-  }
-
-  @Test
-  void shouldNotReuseADeletedVersionNumber() {
-    // given
-    final var form = sampleFormRecord();
-    formCreatedApplier.applyState(form.getFormKey(), form);
-    final var formV2 = sampleFormRecord(2, 2L);
-    formCreatedApplier.applyState(formV2.getFormKey(), formV2);
-
-    // when
-    formDeletedApplier.applyState(form.getFormKey(), form);
-    formDeletedApplier.applyState(formV2.getFormKey(), formV2);
-
-    // then
-    final var nextAvailableVersion =
-        formState.getNextFormVersion(form.getFormId(), form.getTenantId());
-    assertThat(nextAvailableVersion).isEqualTo(3);
-  }
-
-  private FormRecord sampleFormRecord(final int version, final long key) {
-    return new FormRecord()
-        .setFormId("form-id")
-        .setVersion(version)
-        .setFormKey(key)
-        .setResourceName("form-1.form")
-        .setChecksum(wrapString("checksum"))
-        .setResource(wrapString("form-resource"))
-        .setTenantId(tenantId);
-  }
-
-  private FormRecord sampleFormRecord() {
-    return sampleFormRecord(1, 1L);
   }
 }
